requirements.txt
asgiref==3.5.2
Django==4.1.3
django-environ==0.9.0
django-social-share==2.3.0
django-tinymce==3.5.0
Pillow==9.3.0
psycopg2-binary==2.9.5
six==1.16.0
sqlparse==0.4.3
